requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
google-api-core==2.4.0
google-api-python-client==2.34.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
pip==21.3.1
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dotenv==0.19.2
regex==2021.11.10
requests==2.27.1
rsa==4.8
setuptools==60.2.0
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.8
wheel==0.37.1
